+++ crates/adapters/src/controller/error.rs
@@ -124,6 +124,11 @@ pub enum ConfigError {
         endpoint_name: String,
         error: String,
     },
+
+    InvalidOutputBufferConfig {
+        endpoint_name: String,
+        error: String,
+    },
 }
 
 impl StdError for ConfigError {}
@@ -151,6 +156,7 @@ impl DetailedError for ConfigError {
             Self::InvalidEncoderConfig { .. } => Cow::from("InvalidEncoderConfig"),
             Self::InvalidParserConfig { .. } => Cow::from("InvalidParserConfig"),
             Self::InvalidTransportConfig { .. } => Cow::from("InvalidTransportConfig"),
+            Self::InvalidOutputBufferConfig { .. } => Cow::from("InvalidOutputBufferConfig"),
         }
     }
 }
@@ -287,6 +293,15 @@ impl Display for ConfigError {
                 write!(
                     f,
                     "invalid transport configuration for endpoint '{endpoint_name}': {error}"
+                )
+            }
+            Self::InvalidOutputBufferConfig {
+                endpoint_name,
+                error,
+            } => {
+                write!(
+                    f,
+                    "invalid output buffer configuration for endpoint '{endpoint_name}': {error}"
                 )
             }
         }
@@ -433,6 +448,13 @@ impl ConfigError {
 
     pub fn invalid_transport_configuration(endpoint_name: &str, error: &str) -> Self {
         Self::InvalidTransportConfig {
+            endpoint_name: endpoint_name.to_string(),
+            error: error.to_string(),
+        }
+    }
+
+    pub fn invalid_output_buffer_configuration(endpoint_name: &str, error: &str) -> Self {
+        Self::InvalidOutputBufferConfig {
             endpoint_name: endpoint_name.to_string(),
             error: error.to_string(),
         }
@@ -868,6 +890,12 @@ impl ControllerError {
     pub fn invalid_transport_configuration(endpoint_name: &str, error: &str) -> Self {
         Self::Config {
             config_error: ConfigError::invalid_transport_configuration(endpoint_name, error),
+        }
+    }
+
+    pub fn invalid_output_buffer_configuration(endpoint_name: &str, error: &str) -> Self {
+        Self::Config {
+            config_error: ConfigError::invalid_output_buffer_configuration(endpoint_name, error),
         }
     }