+++ python/tests/test_wireframes.py
@@ -1,7 +1,11 @@
+import time
 import unittest
 import pandas as pd
+from kafka import KafkaProducer, KafkaConsumer, TopicPartition
+from kafka.admin import KafkaAdminClient, NewTopic
 
 from feldera import SQLContext, SQLSchema
+from feldera.formats import JSONFormat, JSONUpdateFormat
 from tests import TEST_CLIENT
 
 
@@ -39,7 +43,7 @@ class TestWireframes(unittest.TestCase):
 
     def test_two_SQLContexts(self):
         # https://github.com/feldera/feldera/issues/1770
-        
+
         sql = SQLContext('sql_context1', TEST_CLIENT).get_or_create()
         sql2 = SQLContext('sql_context2', TEST_CLIENT).get_or_create()
 
@@ -104,6 +108,178 @@ class TestWireframes(unittest.TestCase):
 
         sql.run_to_completion()
 
+    def test_df_without_columns(self):
+
+        sql = SQLContext('df_without_columns', TEST_CLIENT).get_or_create()
+        TBL_NAME = "student"
+
+        df = pd.DataFrame([(1, "a"), (2, "b"), (3, "c")])
+
+        sql.register_table(TBL_NAME, SQLSchema({"id": "INT", "name": "STRING"}))
+        sql.register_view("s", f"SELECT * FROM {TBL_NAME}")
+
+        with self.assertRaises(ValueError):
+            sql.connect_source_pandas(TBL_NAME, df)
+
+    def test_sql_error(self):
+        sql = SQLContext('sql_error', TEST_CLIENT).get_or_create()
+        TBL_NAME = "student"
+        df = pd.DataFrame([(1, "a"), (2, "b"), (3, "c")], columns=["id", "name"])
+        sql.register_table(TBL_NAME, SQLSchema({"id": "INT", "name": "STRING"}))
+        sql.register_view("s", f"SELECT FROM blah")
+        sql.connect_source_pandas(TBL_NAME, df)
+        _ = sql.listen("s")
+
+        with self.assertRaises(Exception):
+            sql.run_to_completion()
+
+    def test_kafka(self):
+        import json
+
+        KAFKA_SERVER = "localhost:19092"
+
+        print("(Re-)creating topics...")
+        admin_client = KafkaAdminClient(
+            bootstrap_servers=KAFKA_SERVER,
+            client_id="test_client"
+        )
+
+        INPUT_TOPIC = "simple_count_input"
+        OUTPUT_TOPIC = "simple_count_output"
+
+        existing_topics = set(admin_client.list_topics())
+        if INPUT_TOPIC in existing_topics:
+            admin_client.delete_topics([INPUT_TOPIC])
+        if OUTPUT_TOPIC in existing_topics:
+            admin_client.delete_topics([OUTPUT_TOPIC])
+        admin_client.create_topics([
+            NewTopic(INPUT_TOPIC, num_partitions=1, replication_factor=1),
+            NewTopic(OUTPUT_TOPIC, num_partitions=1, replication_factor=1),
+        ])
+        print("Topics ready")
+
+        # Produce rows into the input topic
+        print("Producing rows into input topic...")
+        num_rows = 1000
+        producer = KafkaProducer(
+            bootstrap_servers=KAFKA_SERVER,
+            client_id="test_client",
+            value_serializer=lambda v: json.dumps(v).encode("utf-8"),
+        )
+        for i in range(num_rows):
+            producer.send("simple_count_input", value={"insert": {"id": i}})
+        print("Input topic contains data")
+
+        TABLE_NAME = "example"
+        VIEW_NAME = "example_count"
+
+        sql = SQLContext('kafka_test', TEST_CLIENT).get_or_create()
+        sql.register_table(TABLE_NAME, SQLSchema({"id": "INT NOT NULL PRIMARY KEY"}))
+        sql.register_view(VIEW_NAME, f"SELECT COUNT(*) as num_rows FROM {TABLE_NAME}")
+
+        PIPELINE_TO_KAFKA_SERVER = "redpanda:9092"
+
+        source_config = {
+            "topics": [INPUT_TOPIC],
+            "bootstrap.servers": PIPELINE_TO_KAFKA_SERVER,
+            "auto.offset.reset": "earliest",
+        }
+
+        kafka_format = JSONFormat().with_update_format(JSONUpdateFormat.InsertDelete).with_array(False)
+
+        sink_config = {
+            "topic": OUTPUT_TOPIC,
+            "bootstrap.servers": PIPELINE_TO_KAFKA_SERVER,
+            "auto.offset.reset": "earliest",
+        }
+
+        sql.connect_source_kafka(TABLE_NAME, "kafka_conn_in", source_config, kafka_format)
+        sql.connect_sink_kafka(VIEW_NAME, "kafka_conn_out", sink_config, kafka_format)
+
+        out = sql.listen(VIEW_NAME)
+        sql.start()
+        time.sleep(10)
+        sql.shutdown()
+        df = out.to_pandas()
+        assert df.shape[0] != 0
+
+    def test_http_get(self):
+        sql = SQLContext("test_http_get", TEST_CLIENT).get_or_create()
+
+        TBL_NAME = "items"
+        VIEW_NAME = "s"
+
+        sql.register_table(TBL_NAME, SQLSchema({"id": "INT", "name": "STRING"}))
+
+        sql.register_view(VIEW_NAME, f"SELECT * FROM {TBL_NAME}")
+
+        path = "https://feldera-basics-tutorial.s3.amazonaws.com/part.json"
+
+        fmt = JSONFormat().with_update_format(JSONUpdateFormat.InsertDelete).with_array(False)
+        sql.connect_source_url(TBL_NAME, "part", path, fmt)
+
+        out = sql.listen(VIEW_NAME)
+
+        sql.run_to_completion()
+
+        df = out.to_pandas()
+
+        assert df.shape[0] == 3
+
+    def test_avro_format(self):
+        from feldera.formats import AvroFormat
+
+        KAFKA_URL_FROM_PIPELINE = "redpanda:9092"
+        KAFKA_SERVER = "localhost:19092"
+        TOPIC = "test_avro_format"
+
+        admin_client = KafkaAdminClient(
+            bootstrap_servers=KAFKA_SERVER,
+            client_id="test_client"
+        )
+        existing_topics = set(admin_client.list_topics())
+        if TOPIC in existing_topics:
+            admin_client.delete_topics([TOPIC])
+
+        df = pd.DataFrame({"id": [1, 2, 3], "name": ["a", "b", "c"]})
+
+        sql = SQLContext("test_avro_format", TEST_CLIENT).get_or_create()
+
+        TBL_NAME = "items"
+        VIEW_NAME = "s"
+
+        sql.register_table(TBL_NAME, SQLSchema({"id": "INT", "name": "STRING"}))
+        sql.register_view(VIEW_NAME, f"SELECT * FROM {TBL_NAME}")
+
+        sql.connect_source_pandas(TBL_NAME, df)
+
+        sink_config = {
+            "topic": TOPIC,
+            "bootstrap.servers": KAFKA_URL_FROM_PIPELINE,
+            "auto.offset.reset": "earliest",
+        }
+
+        avro_format = AvroFormat().with_schema({
+            "type": "record",
+            "name": "items",
+            "fields": [
+                {"name": "id", "type": ["null", "int"]},
+                {"name": "name", "type": ["null", "string"]}
+            ]
+        })
+
+        sql.connect_sink_kafka(VIEW_NAME, "out_avro_kafka", sink_config, avro_format)
+
+        sql.run_to_completion()
+
+        consumer = KafkaConsumer(
+            TOPIC,
+            bootstrap_servers=KAFKA_SERVER,
+            auto_offset_reset='earliest',
+        )
+
+        msg = next(consumer)
+        assert msg.value is not None
 
 if __name__ == '__main__':
     unittest.main()